Comm system

+++ src/RSMixer.cpp
@@ -13,7 +13,7 @@ RSMixer::~RSMixer() { Mix_Quit(); }
 
 void RSMixer::PlayMusic(uint32_t index) {
     MemMusic *mus = this->music->GetMusic(index);
-    SDL_RWops *rw = SDL_RWFromConstMem(mus->data, mus->size);
+    SDL_RWops *rw = SDL_RWFromConstMem(mus->data, (int) mus->size);
     Mix_Music *music = Mix_LoadMUSType_RW(rw, MUS_MID, 1);
     int error = Mix_PlayMusic(music, 1);
     if (error == -1) {
